+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Library/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Library/releases)
 
-- Added: Character and realm utility function with separate return values
+- Added: Optional CurseForge link button for the About section
+- Changed: About section now determines game version and game flavor from the library